+++ core/components/backupmodx/docs/changelog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2021-12-06
+
+### Added
+
+- PHP 8 compatibility
+- Install the composer dependencies directly on the server
+
+### Changed
+
+- Change dependency from chumper/zipper to madnest/madzipper
+
 ## [3.0.5] - 2021-12-06
 
 ### Changed